+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
+- Action Events into ActionSpriteTag
 - DeviceInputChangedListener component
 
 ## [2.0.0] - 2026-03-20